hooked up functionality of diaglog buttons
